discretesignal now converts bools to float64

+++ src/signal.jl
@@ -3,7 +3,7 @@
 # convert bool to int and squish a continuous stream to unique, discrete occurances
 
 function discretesignal(ba::TimeArray{Bool,1})
-    iba = int(ba.values)
+    iba = float64(ba.values)
     idx = [1]
     for b in 2:length(ba)
         if ba.values[b] !== ba.values[b-1]
